code/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Giovanni Spadaccini <giospadaccini74@gmail.com>, Angelo Huang <huangelo02@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
pyyaml = "^6.0.1"

[tool.poetry.scripts]
start = 'backend:start'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
